fixed: unify([:x,:y],:(P(f(x),x)),:(P(y,a))) makes [a,f(x)] not [a,f(a)]
This is caused by parallel unify but not marged them.

I solved it by introducint fixed point of substitution.
if sigma == sigma * sigma, the sigma is the fixed point.

In Chang&Lee's TPU, an unification everytime searchs the
difference from the start of terms, and the difference become a
substitution for a variable. This algorithm doesn't make my problem.

Fixed point is an alternative solution for it. I beleave so.

	modified:   reso.jl
	modified:   testreso.jl